chore: move type file to src->types folder
